Rename SamCliCommand -> SamCliInvocation.
