+++ src/shared/sam/cli/samCliInfo.ts
@@ -9,42 +9,10 @@ import { extensionSettingsPrefix } from '../../constants'
 import { DefaultSettingsConfiguration } from '../../settingsConfiguration'
 import { ChildProcess, ChildProcessResult } from '../../utilities/childProcess'
 import { SamCliConfiguration } from './samCliConfiguration'
+import { SamCliInvocation } from './samCliInvocation'
 import { DefaultSamCliLocationProvider } from './samCliLocator'
 
-/**
- * Represents a call to sam cli
- * Callers are expected to ensure SAM CLI is installed and has been configured
- */
-export abstract class SamCliCommand<T> {
-
-    private readonly _samCliLocation: string | undefined
-
-    protected constructor(config: SamCliConfiguration) {
-        this._samCliLocation = config.getSamCliLocation()
-    }
-
-    public abstract execute(): Thenable<T>
-
-    protected get samCliLocation(): string {
-        if (!this._samCliLocation) {
-            throw new Error('SAM CLI location not configured')
-        }
-
-        return this._samCliLocation
-    }
-
-    /**
-     * Ensures the command is properly set up to run, throws Error if not.
-     * Derived classes should likely call validate at the start of their execute implementations.
-     */
-    protected async validate(): Promise<void> {
-        if (!this._samCliLocation) {
-            throw new Error('SAM CLI location not configured')
-        }
-    }
-}
-
-export class SamCliInfoCommand extends SamCliCommand<SamCliInfoResponse> {
+export class SamCliInfoInvocation extends SamCliInvocation<SamCliInfoResponse> {
 
     public constructor(config: SamCliConfiguration = new SamCliConfiguration(
         new DefaultSettingsConfiguration(extensionSettingsPrefix),
