+++ src/test/shared/sam/cli/samCliInfo.test.ts
@@ -6,12 +6,12 @@
 'use strict'
 
 import * as assert from 'assert'
-import { SamCliInfoCommand, SamCliInfoResponse } from '../../../../shared/sam/cli/samCliCommand'
 import { SamCliConfiguration } from '../../../../shared/sam/cli/samCliConfiguration'
+import { SamCliInfoInvocation, SamCliInfoResponse } from '../../../../shared/sam/cli/samCliInfo'
 
 describe('SamInfoCliCommand', async () => {
 
-    class TestSamCliInfoCommand extends SamCliInfoCommand {
+    class TestSamCliInfoCommand extends SamCliInfoInvocation {
         public convertOutput(text: string): SamCliInfoResponse | undefined {
             return super.convertOutput(text)
         }
@@ -22,7 +22,7 @@ describe('SamInfoCliCommand', async () => {
             getSamCliLocation: () => undefined
         } as any as SamCliConfiguration
 
-        const command = new SamCliInfoCommand(samCliConfig)
+        const command = new SamCliInfoInvocation(samCliConfig)
 
         try {
             await command.execute()
